Rename observables.ts to asyncMap.ts to reflect sole export.
